refactor(app): extract question answer-editing factory from dock
Move the answer-draft mutations (customUpdate/picked/pick/toggle/
customToggle/customOpen/commitCustom) out of session-question-dock.tsx
into createQuestionAnswerEditing, a deps-injected factory in
question-answer-editing.ts. focus is injected as a callback so the
keyboard-nav slice can later own it without a back-import. The shared
live-store type QuestionStore moves to question-draft.ts (cache becomes
its persisted subset). selectOption stays in the dock as a thin
dispatcher over the factory primitives. Pure extraction, no behavior
change; options().length becomes the injected optionCount().
